backend/test-with-python/requirements.txt
pytest
mysql-connector-python
cryptography
pytest-html
jinja2
pymongo
redis
postgresql